pyproject.toml
[tool.poetry]
name = "check-stock-price"
version = "0.1.0"
description = ""
authors = ["Chris Birch <datapointchris@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "check_stock_price"}]

[tool.poetry.scripts]
check-stock-price = "check_stock_price.main:app"

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.9.0"
requests = "^2.31.0"
rich = "^13.5.2"
python-dotenv = "^1.0.0"
pendulum = "^2.1.2"
pandas = "^2.1.0"
seaborn = "^0.12.2"



[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
